[Parley] Refactor: Extract GffFieldFactory from DialogWriter (#534)
Extracted field creation methods into GffFieldFactory class:
- BuildLocStringFieldData, BuildCExoStringFieldData, BuildCResRefFieldData
- AddLabelAndField, GetOrCreateTextOffset
- CalculateEntryFieldCount, CalculateReplyFieldCount

DialogWriter.cs: 1,915 → 1,767 lines (-148)
GffFieldFactory.cs: 201 lines (new)

All 500 tests pass ✅
